requirements/base.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# API Clients
google-cloud-translate==3.14.1
openai==1.3.8

# Local GPU/Translation
transformers==4.35.2
torch==2.1.2
accelerate==0.25.0
safetensors==0.4.1

# Caching & Queue
redis==5.0.1
celery==5.3.4

# Logging & Monitoring
python-json-logger==2.0.7
